Deletes example ssh project, moves code to opkssh
