+++ opkssh/main_test.go
@@ -17,83 +17,9 @@
 package main
 
 import (
-	"context"
 	"errors"
-	"strings"
 	"testing"
-
-	"github.com/lestrrat-go/jwx/v2/jwa"
-	"github.com/openpubkey/openpubkey/client"
-	"github.com/openpubkey/openpubkey/examples/ssh/sshcert"
-	"github.com/openpubkey/openpubkey/pktoken"
-	"github.com/openpubkey/openpubkey/providers"
-	"github.com/stretchr/testify/require"
-
-	"github.com/openpubkey/openpubkey/util"
-	"golang.org/x/crypto/ssh"
 )
-
-func AllowAllPolicyEnforcer(userDesired string, pkt *pktoken.PKToken) error {
-	return nil
-}
-
-func TestSshCli(t *testing.T) {
-	providerOpts := providers.DefaultMockProviderOpts()
-	op, _, _, err := providers.NewMockProvider(providerOpts)
-	require.NoError(t, err)
-
-	certBytes, seckeySshPem, err := Login(op)
-	require.NoError(t, err)
-	require.NotNil(t, certBytes)
-	require.NotNil(t, seckeySshPem)
-}
-
-func TestAuthorizedKeysCommand(t *testing.T) {
-	alg := jwa.ES256
-	signer, err := util.GenKeyPair(alg)
-	require.NoError(t, err)
-
-	providerOpts := providers.DefaultMockProviderOpts()
-	op, _, idtTemplate, err := providers.NewMockProvider(providerOpts)
-	require.NoError(t, err)
-
-	mockEmail := "arthur.aardvark@example.com"
-	idtTemplate.ExtraClaims = map[string]any{
-		"email": mockEmail,
-	}
-
-	client, err := client.New(op, client.WithSigner(signer, alg))
-	require.NoError(t, err)
-
-	pkt, err := client.Auth(context.Background())
-	require.NoError(t, err)
-
-	principals := []string{"guest", "dev"}
-	cert, err := sshcert.New(pkt, principals)
-	require.NoError(t, err)
-
-	sshSigner, err := ssh.NewSignerFromSigner(signer)
-	require.NoError(t, err)
-
-	signerMas, err := ssh.NewSignerWithAlgorithms(sshSigner.(ssh.AlgorithmSigner),
-		[]string{ssh.KeyAlgoECDSA256})
-	require.NoError(t, err)
-
-	sshCert, err := cert.SignCert(signerMas)
-	require.NoError(t, err)
-
-	certTypeAndCertB64 := ssh.MarshalAuthorizedKey(sshCert)
-	typeArg := strings.Split(string(certTypeAndCertB64), " ")[0]
-	certB64Arg := strings.Split(string(certTypeAndCertB64), " ")[1]
-
-	userArg := "user"
-	pubkeyList, err := authorizedKeysCommand(userArg, typeArg, certB64Arg, AllowAllPolicyEnforcer, op)
-	require.NoError(t, err)
-
-	expectedPubkeyList := "cert-authority ecdsa-sha2-nistp256"
-	require.Contains(t, pubkeyList, expectedPubkeyList)
-
-}
 
 func TestIsOpenSSHVersion8Dot1OrGreater(t *testing.T) {
 	tests := []struct {
